Mouse Gestures completely refactored
Now it doesn't listen to mousemove and wheel when it doesn't need to.
Also many code improvements. Previous version had no issues (that I know), but it was a mess.

Also, now it's possible for an ucJS to break into each WebExtension context and run code. I do this in Mouse Gestures.


Upstream-Commit: 23ac9f300847b5d9594a3c39dc3e789d6d9ad26a